Add comparison class to game flow class
